requirements.txt
# core
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-Login==0.5.0
Flask-WTF==0.15.1
python-dotenv==0.19.0
Werkzeug==2.0.1
SQLAlchemy==1.4.23
alembic==1.7.1

# DB
flask_migrate==4.1.0
flask_sqlalchemy==3.1.1
sqlalchemy==2.0.38

# tools
flask_login==0.6.3
flask-dance==7.1.0
celery>=5.3.0
redis>=4.5.0
colorama==0.4.6
PyJWT~=2.10.1
WTForms-Alchemy==0.19.0

# utils
email_validator==2.2.0
blinker==1.9.0

# env
python-dotenv==1.0.1

# deployment
gunicorn==23.0.0
Flask-Minify==0.49
Flask-CDN==1.5.3

flask_mysqldb
psycopg2-binary
mysql-connector-python
mysqlclient
boto3==1.34.69
flasgger

APScheduler>=3.10.0
requests
psutil
python-decouple
croniter
logger
cryptography
paramiko==3.4.0

flower>=2.0.0  # For Celery monitoring

flask_celery